Time: 6 ms (44.90%) | Memory: 23.9 MB (59.66%) - LeetSync

+++ 56-merge-intervals/merge-intervals.cpp
@@ -5,7 +5,8 @@ public:
 
         vector<vector<int>> ans;
         ans.push_back(intervals[0]);
-        for(int i = 1; i<intervals.size(); ++i){
+
+        for(int i = 1; i< (int)intervals.size(); ++i){
             if(ans.back()[1] >= intervals[i][0]){
                 ans.back()[1] = max(ans.back()[1], intervals[i][1]);
             }
